+++ command.cpp
@@ -129,11 +129,11 @@ void static position(std::istringstream& is, Position & pos){
 */
 void static doPerft(unsigned int n, Position & pos){
 	std::string token;
-	unsigned long elapsed = std::chrono::duration_cast<std::chrono::milliseconds >(std::chrono::steady_clock::now().time_since_epoch()).count();
+	unsigned long long elapsed = std::chrono::duration_cast<std::chrono::milliseconds >(std::chrono::steady_clock::now().time_since_epoch()).count();
 	unsigned long long res=pos.perft(n);
 	elapsed = std::chrono::duration_cast<std::chrono::milliseconds >(std::chrono::steady_clock::now().time_since_epoch()).count()-elapsed;
 	sync_cout<<"Perft "<<n<<" leaf nodes: "<<res<<sync_endl;
-	sync_cout<<elapsed<<"ms "<<((double)res)/elapsed<<" kN/s"<<sync_endl;
+	sync_cout<<elapsed<<"ms "<<((double)res)/(double)elapsed<<" kN/s"<<sync_endl;
 }
 
 
@@ -345,7 +345,7 @@ void uciLoop(){
 			doPerft(stoi(token), pos);
 		}
 		else if (token =="divide" && (is>>token)){
-			unsigned long res=pos.divide(stoi(token));
+			unsigned long long res=pos.divide(stoi(token));
 			sync_cout<<"divide Res="<<res<<sync_endl;
 		}
 		else if (token =="go"){
